Move net::ip tests to their own files
